requirements.txt
gunicorn==23.0.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
msgpack==1.1.0
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
uvicorn==0.24.0